Refactor studios controllers to reduce code duplication
- Extract shared studios controller logic into studiosController.js
- Update moviestudios.js and tvstudios.js to use shared controller
- Reduces duplication from 60% to near zero for moviestudios.js
- Add comment explaining acceptable duplication in moviesrecommended.js tab management
